# src/docs/swagger.yaml
openapi: 3.1.0
info:
  title: Test task API
  version: 1.0.0
  description: API for Test task
servers:
  - url: http://localhost:7852
paths:
...
  /api/tasks:
    post:
      summary: Create Task
      tags:
        - Tasks
      operationId: Create Task
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequestBody"
      responses:
        200:
          description: Create Task
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Task create successfully
    get:
      summary: Tasks list
      tags:
        - Tasks
      operationId: Tasks list
      security:
        - BearerAuth: [ ]
      responses:
        200:
          description: Tasks list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Tasks list
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskResponseBody"
                    count:
                      type: number
...
components:
  schemas:
    TaskRequestBody:
      type: object
      properties:
        assignee_id:
          type: string
          format: uuid
        title:
          type: string
          required: true
          default: test
        description:
          type: string
          required: false
          default: test description
        priority:
          type: number
          enum: [ 1, 2, 3, 4, 5 ]
          required: false
          default: 3
        status:
          type: string
          enum: [ TO_DO, In_Progress, DONE ]
          required: false
          default: TO_DO
        due_date:
          type: string
          format: date-time
          required: true
          default: 2024-12-12
        time_tracking:
          type: number
          required: false
          default: 2
    TaskResponseBody:
      type: object
      properties:
        assignee_id:
          type: string
          format: uuid
        reporter_id:
          type: string
          format: uuid
        title:
          type: string
          required: true
          default: test
        description:
          type: string
          required: false
          default: test description
        priority:
          type: number
          enum: [ 1, 2, 3, 4, 5 ]
          required: false
          default: 3
        status:
          type: string
          enum: [ TO_DO, In_Progress, DONE ]
          required: false
          default: TO_DO
        due_date:
          type: string
          format: date-time
          required: true
          default: 2024-12-12
        time_tracking:
          type: number
          required: false
          default: 2
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT